audion_signal - BUG - fixed bug in method to_wav
Former-commit-id: 8cd0a98286113c21bd1a03547c376d088c809a01
Former-commit-id: e66c9e25b0f232f5b69a47cd2fb3f8ebdb04e6f1

+++ sound_classification/audio_signal.py
@@ -24,7 +24,7 @@ class AudioSignal:
         return cls(rate, data, path[path.rfind('/')+1:])
 
     def to_wav(self, path):
-        wave.write(filename=path, rate=self.rate, data=self.data)
+        wave.write(filename=path, rate=int(self.rate), data=self.data.astype(dtype=np.int16))
 
     def empty(self):
         return len(self.data) == 0
